+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 Phase 3 release.
 
+### Added
+
+- Silly (but nice) ASCII art banner for help command.
+- Default creds/non-creds are now config files created on `cloudsmith token`.
+
 ## [0.2.2] - 2017-12-03
 
 ### Fixed